+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Allow search by barcode. Refs UIU-1708.
 * Work with multiple and single `refund` action. Refs UIU-1897.
 * Show the number of open requests in `LoanActionDialog`. Refs UIU-1890.
+* Remove optional `action` fields if no value was setted. Refs UIU-1773.
 * Show the number of open requests in the Claim returned bulk action modal. Refs UIU-1891.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)